+++ frontend/spotted-apple/data-access/types.tsx
