+++ frontend/spotted-apple/lib/forms/components.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { useFormStatus } from 'react-dom'
-import { InputField, ButtonDisplay, InputFieldsErrors} from '@/app/lib/forms/types'
+import { InputField, ButtonDisplay, InputFieldsErrors} from '@/lib/forms/types'
 
 export function FormField({id, placeHolder, initialValue, onChange}: InputField) {
     return (<input
