+++ frontend/spotted-apple/lib/styles/globals.css
